requirements.txt
pyyaml>=6.0
tqdm>=4.65.0
colorama>=0.4.6
